Fixed Incorrect Autofill when Adding/Editing Users
Previously when editing for adding a user, the login credential fields were being auto completed to match the current logged in users login details. Now when creating a user it leaves these fields empty as it should. When editing a user it will now also auto fill to the correct login name.